Move flag provider to cfg
